+++ modules/src/ics03_connection/context.rs
@@ -22,9 +22,8 @@ pub trait ConnectionReader {
     /// Returns the current height of the local chain.
     fn host_current_height(&self) -> Height;
 
-    /// Returns the number of consensus state entries that the local chain maintains. The history
-    /// size determines the pruning window of the host chain.
-    fn host_chain_history_size(&self) -> usize;
+    /// Returns the oldest height available on the local chain.
+    fn host_oldest_height(&self) -> Height;
 
     /// Returns the prefix that the local chain uses in the KV store.
     fn commitment_prefix(&self) -> CommitmentPrefix;