+++ modules/src/mock/context.rs
@@ -102,10 +102,10 @@ pub struct MockContext {
 impl Default for MockContext {
     fn default() -> Self {
         Self::new(
-            ChainId::new("mockgaia".to_string(), 1),
+            ChainId::new("mockgaia".to_string(), 0),
             HostType::Mock,
             5,
-            Height::new(1, 5),
+            Height::new(0, 5),
         )
     }
 }
@@ -450,9 +450,9 @@ impl ConnectionReader for MockContext {
         self.latest_height
     }
 
-    /// Returns the number of consensus state historical entries for the local chain.
-    fn host_chain_history_size(&self) -> usize {
-        self.max_history_size
+    fn host_oldest_height(&self) -> Height {
+        // history must be non-empty, so `self.history[0]` is valid
+        self.history[0].height()
     }
 
     fn commitment_prefix(&self) -> CommitmentPrefix {